feat: review changes

+++ tests/test_network.py
@@ -1,4 +1,5 @@
 import pytest
+from decimal import Decimal
 from web3 import Web3
 
 from wallet.network.provider import get_provider, is_connected, get_eth_balance, get_gas_price
@@ -14,10 +15,10 @@ def test_get_eth_balance():
     # vitalik.eth public address, used as a safe read-only test vector
     address = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
     balance = get_eth_balance(address)
-    assert isinstance(balance, float)
+    assert isinstance(balance, Decimal)
     assert balance >= 0
 
 def test_get_gas_price():
     gas_price = get_gas_price()
-    assert isinstance(gas_price, float)
+    assert isinstance(gas_price, Decimal)
     assert gas_price > 0
